Moved style constext parsing to own file
